requirements.txt
Django~=2.2.0
djangorestframework~=3.11

# tests
pytest==5.2.1
pytest-django==3.5.1